[embeddable] remove embeddable factory methods from setup and start API (#204797)
Part of embeddable rebuild cleanup. PR removes legacy embeddable factory
registration APIs.

---------

(cherry picked from commit 3b61e7bea7408e586faa4be6464e963b50385896)
