+++ src/plugins/embeddable/public/kibana_services.ts
@@ -11,7 +11,7 @@ import { BehaviorSubject } from 'rxjs';
 
 import { CoreStart } from '@kbn/core/public';
 
-import { EmbeddableStart, EmbeddableStartDependencies } from '.';
+import { EmbeddableStart, EmbeddableStartDependencies } from './types';
 
 export let core: CoreStart;
 export let embeddableStart: EmbeddableStart;
