+++ src/plugins/embeddable/public/lib/embeddables/index.ts
@@ -8,10 +8,7 @@
  */
 
 export * from '../../../common/lib/saved_object_embeddable';
-export * from './default_embeddable_factory_provider';
 export { Embeddable } from './embeddable';
 export { EmbeddableErrorHandler } from './embeddable_error_handler';
-export * from './embeddable_factory';
-export * from './embeddable_factory_definition';
 export { ErrorEmbeddable } from './error_embeddable';
 export type { EmbeddableInput, EmbeddableOutput, IEmbeddable } from './i_embeddable';
