+++ src/plugins/embeddable/public/plugin.test.ts
@@ -10,104 +10,6 @@
 import { coreMock } from '@kbn/core/public/mocks';
 import { testPlugin } from './tests/test_plugin';
 
-describe('embeddable factory', () => {
-  const coreSetup = coreMock.createSetup();
-  const coreStart = coreMock.createStart();
-  const { setup, doStart } = testPlugin(coreSetup, coreStart);
-  const start = doStart();
-  const embeddableFactoryId = 'ID';
-  const embeddableFactory = {
-    type: embeddableFactoryId,
-    create: jest.fn(),
-    getDisplayName: () => 'Test',
-    isEditable: () => Promise.resolve(true),
-    extract: jest.fn().mockImplementation((state) => ({ state, references: [] })),
-    inject: jest.fn().mockImplementation((state) => state),
-    telemetry: jest.fn().mockResolvedValue({}),
-    latestVersion: '7.11.0',
-    migrations: { '7.11.0': jest.fn().mockImplementation((state) => state) },
-  } as any;
-  const embeddableState = {
-    id: embeddableFactoryId,
-    type: embeddableFactoryId,
-    my: 'state',
-  } as any;
-
-  const containerEmbeddableFactoryId = 'CONTAINER';
-  const containerEmbeddableFactory = {
-    type: containerEmbeddableFactoryId,
-    latestVersion: '1.0.0',
-    create: jest.fn(),
-    getDisplayName: () => 'Container',
-    isContainer: true,
-    isEditable: () => Promise.resolve(true),
-    extract: jest.fn().mockImplementation((state) => ({ state, references: [] })),
-    inject: jest.fn().mockImplementation((state) => state),
-    telemetry: jest.fn().mockResolvedValue({}),
-    migrations: { '7.12.0': jest.fn().mockImplementation((state) => state) },
-  };
-
-  const containerState = {
-    id: containerEmbeddableFactoryId,
-    type: containerEmbeddableFactoryId,
-    some: 'state',
-    panels: [
-      {
-        ...embeddableState,
-      },
-    ],
-  } as any;
-
-  setup.registerEmbeddableFactory(embeddableFactoryId, embeddableFactory);
-  setup.registerEmbeddableFactory(containerEmbeddableFactoryId, containerEmbeddableFactory);
-
-  test('cannot register embeddable factory with the same ID', async () => {
-    expect(() =>
-      setup.registerEmbeddableFactory(embeddableFactoryId, embeddableFactory)
-    ).toThrowError(
-      'Embeddable factory [embeddableFactoryId = ID] already registered in Embeddables API.'
-    );
-  });
-
-  test('embeddableFactory extract function gets called when calling embeddable extract', () => {
-    start.extract(embeddableState);
-    expect(embeddableFactory.extract).toBeCalledWith(embeddableState);
-  });
-
-  test('embeddableFactory inject function gets called when calling embeddable inject', () => {
-    start.inject(embeddableState, []);
-    expect(embeddableFactory.extract).toBeCalledWith(embeddableState);
-  });
-
-  test('embeddableFactory telemetry function gets called when calling embeddable telemetry', () => {
-    start.telemetry(embeddableState, {});
-    expect(embeddableFactory.telemetry).toBeCalledWith(embeddableState, {});
-  });
-
-  test('embeddableFactory migrate function gets called when calling embeddable migrate', () => {
-    start.getAllMigrations!()['7.11.0']!(embeddableState);
-    expect(embeddableFactory.migrations['7.11.0']).toBeCalledWith(embeddableState);
-  });
-
-  test('panels inside container get automatically migrated when migrating conta1iner', () => {
-    start.getAllMigrations!()['7.11.0']!(containerState);
-    expect(embeddableFactory.migrations['7.11.0']).toBeCalledWith(embeddableState);
-  });
-
-  test('migrateToLatest returns list of all migrations', () => {
-    const migrations = start.getAllMigrations();
-    expect(migrations).toMatchSnapshot();
-  });
-
-  test('migrateToLatest calls correct migrate functions', () => {
-    start.migrateToLatest!({
-      state: embeddableState,
-      version: '7.11.0',
-    });
-    expect(embeddableFactory.migrations['7.11.0']).toBeCalledWith(embeddableState);
-  });
-});
-
 describe('embeddable enhancements', () => {
   const coreSetup = coreMock.createSetup();
   const coreStart = coreMock.createStart();
