+++ src/plugins/embeddable/public/plugin.tsx
@@ -8,10 +8,6 @@
  */
 
 import { Subscription } from 'rxjs';
-import { identity } from 'lodash';
-import type { SerializableRecord } from '@kbn/utility-types';
-import { UiActionsSetup, UiActionsStart } from '@kbn/ui-actions-plugin/public';
-import { Start as InspectorStart } from '@kbn/inspector-plugin/public';
 import {
   PluginInitializerContext,
   CoreSetup,
@@ -20,26 +16,8 @@ import {
   PublicAppInfo,
 } from '@kbn/core/public';
 import { Storage } from '@kbn/kibana-utils-plugin/public';
-import { UsageCollectionStart } from '@kbn/usage-collection-plugin/public';
-import { migrateToLatest, PersistableStateService } from '@kbn/kibana-utils-plugin/common';
-import { SavedObjectsManagementPluginStart } from '@kbn/saved-objects-management-plugin/public';
-import type { ContentManagementPublicStart } from '@kbn/content-management-plugin/public';
-import type { SavedObjectTaggingOssPluginStart } from '@kbn/saved-objects-tagging-oss-plugin/public';
-import {
-  EmbeddableFactoryRegistry,
-  EnhancementsRegistry,
-  EnhancementRegistryDefinition,
-  EnhancementRegistryItem,
-} from './types';
+import { migrateToLatest } from '@kbn/kibana-utils-plugin/common';
 import { bootstrap } from './bootstrap';
-import {
-  EmbeddableFactory,
-  EmbeddableInput,
-  EmbeddableOutput,
-  defaultEmbeddableFactoryProvider,
-  IEmbeddable,
-} from './lib';
-import { EmbeddableFactoryDefinition } from './lib/embeddables/embeddable_factory_definition';
 import { EmbeddableStateTransfer } from './lib/state_transfer';
 import { EmbeddableStateWithType, CommonEmbeddableStartContract } from '../common/types';
 import {
@@ -52,90 +30,19 @@ import { getAllMigrations } from '../common/lib/get_all_migrations';
 import { setKibanaServices } from './kibana_services';
 import { registerReactEmbeddableFactory } from './react_embeddable_system';
 import { registerAddFromLibraryType } from './add_from_library/registry';
+import { EnhancementsRegistry } from './enhancements/registry';
+import {
+  EmbeddableSetup,
+  EmbeddableSetupDependencies,
+  EmbeddableStart,
+  EmbeddableStartDependencies,
+} from './types';
 
-export interface EmbeddableSetupDependencies {
-  uiActions: UiActionsSetup;
-}
-
-export interface EmbeddableStartDependencies {
-  uiActions: UiActionsStart;
-  inspector: InspectorStart;
-  usageCollection: UsageCollectionStart;
-  contentManagement: ContentManagementPublicStart;
-  savedObjectsManagement: SavedObjectsManagementPluginStart;
-  savedObjectsTaggingOss?: SavedObjectTaggingOssPluginStart;
-}
-
-export interface EmbeddableSetup {
-  /**
-   * Register a saved object type with the "Add from library" flyout.
-   *
-   * @example
-   *  registerAddFromLibraryType({
-   *    onAdd: (container, savedObject) => {
-   *      container.addNewPanel({
-   *        panelType: CONTENT_ID,
-   *        initialState: savedObject.attributes,
-   *      });
-   *    },
-   *    savedObjectType: MAP_SAVED_OBJECT_TYPE,
-   *    savedObjectName: i18n.translate('xpack.maps.mapSavedObjectLabel', {
-   *      defaultMessage: 'Map',
-   *    }),
-   *    getIconForSavedObject: () => APP_ICON,
-   *  });
-   */
-  registerAddFromLibraryType: typeof registerAddFromLibraryType;
-
-  /**
-   * Registers an async {@link ReactEmbeddableFactory} getter.
-   */
-  registerReactEmbeddableFactory: typeof registerReactEmbeddableFactory;
-
-  /**
-   * @deprecated use {@link registerReactEmbeddableFactory} instead.
-   */
-  registerEmbeddableFactory: <
-    I extends EmbeddableInput,
-    O extends EmbeddableOutput,
-    E extends IEmbeddable<I, O> = IEmbeddable<I, O>
-  >(
-    id: string,
-    factory: EmbeddableFactoryDefinition<I, O, E>
-  ) => () => EmbeddableFactory<I, O, E>;
-  /**
-   * @deprecated
-   */
-  registerEnhancement: (enhancement: EnhancementRegistryDefinition) => void;
-}
-
-export interface EmbeddableStart extends PersistableStateService<EmbeddableStateWithType> {
-  /**
-   * @deprecated use {@link registerReactEmbeddableFactory} instead.
-   */
-  getEmbeddableFactory: <
-    I extends EmbeddableInput = EmbeddableInput,
-    O extends EmbeddableOutput = EmbeddableOutput,
-    E extends IEmbeddable<I, O> = IEmbeddable<I, O>
-  >(
-    embeddableFactoryId: string
-  ) => EmbeddableFactory<I, O, E> | undefined;
-
-  /**
-   * @deprecated
-   */
-  getEmbeddableFactories: () => IterableIterator<EmbeddableFactory>;
-  getStateTransfer: (storage?: Storage) => EmbeddableStateTransfer;
-}
 export class EmbeddablePublicPlugin implements Plugin<EmbeddableSetup, EmbeddableStart> {
-  private readonly embeddableFactoryDefinitions: Map<string, EmbeddableFactoryDefinition> =
-    new Map();
-  private readonly embeddableFactories: EmbeddableFactoryRegistry = new Map();
-  private readonly enhancements: EnhancementsRegistry = new Map();
   private stateTransferService: EmbeddableStateTransfer = {} as EmbeddableStateTransfer;
-  private isRegistryReady = false;
   private appList?: ReadonlyMap<string, PublicAppInfo>;
   private appListSubscription?: Subscription;
+  private enhancementsRegistry = new EnhancementsRegistry();
 
   constructor(initializerContext: PluginInitializerContext) {}
 
@@ -145,17 +52,11 @@ export class EmbeddablePublicPlugin implements Plugin<EmbeddableSetup, Embeddabl
     return {
       registerReactEmbeddableFactory,
       registerAddFromLibraryType,
-
-      registerEmbeddableFactory: this.registerEmbeddableFactory,
-      registerEnhancement: this.registerEnhancement,
+      registerEnhancement: this.enhancementsRegistry.registerEnhancement,
     };
   }
 
   public start(core: CoreStart, deps: EmbeddableStartDependencies): EmbeddableStart {
-    this.embeddableFactoryDefinitions.forEach((def) => {
-      this.embeddableFactories.set(def.type, defaultEmbeddableFactoryProvider(def));
-    });
-
     this.appListSubscription = core.application.applications$.subscribe((appList) => {
       this.appList = appList;
     });
@@ -165,24 +66,19 @@ export class EmbeddablePublicPlugin implements Plugin<EmbeddableSetup, Embeddabl
       core.application.currentAppId$,
       this.appList
     );
-    this.isRegistryReady = true;
 
     const commonContract: CommonEmbeddableStartContract = {
-      getEmbeddableFactory: this
-        .getEmbeddableFactory as unknown as CommonEmbeddableStartContract['getEmbeddableFactory'],
-      getEnhancement: this.getEnhancement,
+      getEnhancement: this.enhancementsRegistry.getEnhancement,
     };
 
     const getAllMigrationsFn = () =>
       getAllMigrations(
-        Array.from(this.embeddableFactories.values()),
-        Array.from(this.enhancements.values()),
+        [],
+        this.enhancementsRegistry.getEnhancements(),
         getMigrateFunction(commonContract)
       );
 
     const embeddableStart: EmbeddableStart = {
-      getEmbeddableFactory: this.getEmbeddableFactory,
-      getEmbeddableFactories: this.getEmbeddableFactories,
       getStateTransfer: (storage?: Storage) =>
         storage
           ? new EmbeddableStateTransfer(
@@ -210,89 +106,4 @@ export class EmbeddablePublicPlugin implements Plugin<EmbeddableSetup, Embeddabl
       this.appListSubscription.unsubscribe();
     }
   }
-
-  private registerEnhancement = (enhancement: EnhancementRegistryDefinition) => {
-    if (this.enhancements.has(enhancement.id)) {
-      throw new Error(`enhancement with id ${enhancement.id} already exists in the registry`);
-    }
-    this.enhancements.set(enhancement.id, {
-      id: enhancement.id,
-      telemetry: enhancement.telemetry || ((state, stats) => stats),
-      inject: enhancement.inject || identity,
-      extract:
-        enhancement.extract ||
-        ((state: SerializableRecord) => {
-          return { state, references: [] };
-        }),
-      migrations: enhancement.migrations || {},
-    });
-  };
-
-  private getEnhancement = (id: string): EnhancementRegistryItem => {
-    return (
-      this.enhancements.get(id) || {
-        id: 'unknown',
-        telemetry: (state, stats) => stats,
-        inject: identity,
-        extract: (state: SerializableRecord) => {
-          return { state, references: [] };
-        },
-        migrations: {},
-      }
-    );
-  };
-
-  private getEmbeddableFactories = () => {
-    this.ensureFactoriesExist();
-    return this.embeddableFactories.values();
-  };
-
-  private registerEmbeddableFactory = <
-    I extends EmbeddableInput = EmbeddableInput,
-    O extends EmbeddableOutput = EmbeddableOutput,
-    E extends IEmbeddable<I, O> = IEmbeddable<I, O>
-  >(
-    embeddableFactoryId: string,
-    factory: EmbeddableFactoryDefinition<I, O, E>
-  ): (() => EmbeddableFactory<I, O, E>) => {
-    if (this.embeddableFactoryDefinitions.has(embeddableFactoryId)) {
-      throw new Error(
-        `Embeddable factory [embeddableFactoryId = ${embeddableFactoryId}] already registered in Embeddables API.`
-      );
-    }
-    this.embeddableFactoryDefinitions.set(embeddableFactoryId, factory);
-
-    return () => {
-      return this.getEmbeddableFactory(embeddableFactoryId);
-    };
-  };
-
-  private getEmbeddableFactory = <
-    I extends EmbeddableInput = EmbeddableInput,
-    O extends EmbeddableOutput = EmbeddableOutput,
-    E extends IEmbeddable<I, O> = IEmbeddable<I, O>
-  >(
-    embeddableFactoryId: string
-  ): EmbeddableFactory<I, O, E> => {
-    if (!this.isRegistryReady) {
-      throw new Error('Embeddable factories can only be retrieved after setup lifecycle.');
-    }
-    this.ensureFactoryExists(embeddableFactoryId);
-    const factory = this.embeddableFactories.get(embeddableFactoryId);
-
-    return factory as EmbeddableFactory<I, O, E>;
-  };
-
-  // These two functions are only to support legacy plugins registering factories after the start lifecycle.
-  private ensureFactoriesExist = () => {
-    this.embeddableFactoryDefinitions.forEach((def) => this.ensureFactoryExists(def.type));
-  };
-
-  private ensureFactoryExists = (type: string) => {
-    if (!this.embeddableFactories.get(type)) {
-      const def = this.embeddableFactoryDefinitions.get(type);
-      if (!def) return;
-      this.embeddableFactories.set(type, defaultEmbeddableFactoryProvider(def));
-    }
-  };
 }
