+++ x-pack/plugins/canvas/common/lib/embeddable_dataurl.ts
@@ -5,11 +5,6 @@
  * 2.0.
  */
 
-import { type ExplicitInputWithAttributes } from '@kbn/embeddable-plugin/public/lib';
-import { EmbeddableInput } from '../../types';
-
-export const encode = (
-  input: ExplicitInputWithAttributes | Partial<EmbeddableInput> | Readonly<EmbeddableInput>
-) => Buffer.from(JSON.stringify(input)).toString('base64');
+export const encode = (input: object) => Buffer.from(JSON.stringify(input)).toString('base64');
 export const decode = (serializedInput: string) =>
   JSON.parse(Buffer.from(serializedInput, 'base64').toString());
